probMap contem as probabilidades em porcentagem

+++ include/prob-Calc.h
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-void calcProb(unordered_map<char, int> &probMap, string str);
-void printProb(unordered_map<char, int> &probMap);
+void calcProb(unordered_map<char, double> &probMap, string str);
+void printProb(unordered_map<char, double> &probMap);
 
 #endif
